Changed SetText to SetPlaceHolder
Changed text to be place holders as opposed to actual strings that need to be deleted prior to entering in desired information. However, commented out hostname PlaceHolder for now, as it's quicker for testing to simply have scanme.nmap.org as the default. (delete later and replace with SetPlaceHolder that's currently commented out)